+++ DockerDotNet.Core/Models/ContainerThrottlingData.cs
@@ -37,7 +37,7 @@ namespace DockerDotNet.Core.Models
         /// <param name="throttledPeriods">Number of periods when the container hit its throttling limit. </param>
         /// <param name="throttledTime">Aggregated time (in nanoseconds) the container was throttled for. </param>
         [JsonConstructor]
-        public ContainerThrottlingData(Option<int?> periods = default, Option<int?> throttledPeriods = default, Option<int?> throttledTime = default)
+        public ContainerThrottlingData(Option<ulong?> periods = default, Option<ulong?> throttledPeriods = default, Option<ulong?> throttledTime = default)
         {
             PeriodsOption = periods;
             ThrottledPeriodsOption = throttledPeriods;
@@ -52,7 +52,7 @@ namespace DockerDotNet.Core.Models
         /// </summary>
         [JsonIgnore]
         [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-        public Option<int?> PeriodsOption { get; private set; }
+        public Option<ulong?> PeriodsOption { get; private set; }
 
         /// <summary>
         /// Number of periods with throttling active. 
@@ -60,14 +60,14 @@ namespace DockerDotNet.Core.Models
         /// <value>Number of periods with throttling active. </value>
         /* <example>0</example> */
         [JsonPropertyName("periods")]
-        public int? Periods { get { return this.PeriodsOption; } set { this.PeriodsOption = new(value); } }
+        public ulong? Periods { get { return this.PeriodsOption; } set { this.PeriodsOption = new(value); } }
 
         /// <summary>
         /// Used to track the state of ThrottledPeriods
         /// </summary>
         [JsonIgnore]
         [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-        public Option<int?> ThrottledPeriodsOption { get; private set; }
+        public Option<ulong?> ThrottledPeriodsOption { get; private set; }
 
         /// <summary>
         /// Number of periods when the container hit its throttling limit. 
@@ -75,14 +75,14 @@ namespace DockerDotNet.Core.Models
         /// <value>Number of periods when the container hit its throttling limit. </value>
         /* <example>0</example> */
         [JsonPropertyName("throttled_periods")]
-        public int? ThrottledPeriods { get { return this.ThrottledPeriodsOption; } set { this.ThrottledPeriodsOption = new(value); } }
+        public ulong? ThrottledPeriods { get { return this.ThrottledPeriodsOption; } set { this.ThrottledPeriodsOption = new(value); } }
 
         /// <summary>
         /// Used to track the state of ThrottledTime
         /// </summary>
         [JsonIgnore]
         [global::System.ComponentModel.EditorBrowsable(global::System.ComponentModel.EditorBrowsableState.Never)]
-        public Option<int?> ThrottledTimeOption { get; private set; }
+        public Option<ulong?> ThrottledTimeOption { get; private set; }
 
         /// <summary>
         /// Aggregated time (in nanoseconds) the container was throttled for. 
@@ -90,7 +90,7 @@ namespace DockerDotNet.Core.Models
         /// <value>Aggregated time (in nanoseconds) the container was throttled for. </value>
         /* <example>0</example> */
         [JsonPropertyName("throttled_time")]
-        public int? ThrottledTime { get { return this.ThrottledTimeOption; } set { this.ThrottledTimeOption = new(value); } }
+        public ulong? ThrottledTime { get { return this.ThrottledTimeOption; } set { this.ThrottledTimeOption = new(value); } }
 
         /// <summary>
         /// Returns the string presentation of the object
@@ -140,9 +140,9 @@ namespace DockerDotNet.Core.Models
 
             JsonTokenType startingTokenType = utf8JsonReader.TokenType;
 
-            Option<int?> periods = default;
-            Option<int?> throttledPeriods = default;
-            Option<int?> throttledTime = default;
+            Option<ulong?> periods = default;
+            Option<ulong?> throttledPeriods = default;
+            Option<ulong?> throttledTime = default;
 
             while (utf8JsonReader.Read())
             {
@@ -161,15 +161,15 @@ namespace DockerDotNet.Core.Models
                     {
                         case "periods":
                             if (utf8JsonReader.TokenType != JsonTokenType.Null)
-                                periods = new Option<int?>(utf8JsonReader.GetInt32());
+                                periods = new Option<ulong?>(utf8JsonReader.GetUInt64());
                             break;
                         case "throttled_periods":
                             if (utf8JsonReader.TokenType != JsonTokenType.Null)
-                                throttledPeriods = new Option<int?>(utf8JsonReader.GetInt32());
+                                throttledPeriods = new Option<ulong?>(utf8JsonReader.GetUInt64());
                             break;
                         case "throttled_time":
                             if (utf8JsonReader.TokenType != JsonTokenType.Null)
-                                throttledTime = new Option<int?>(utf8JsonReader.GetInt32());
+                                throttledTime = new Option<ulong?>(utf8JsonReader.GetUInt64());
                             break;
                         default:
                             break;
